+++ internal/commands/all/resource.go
@@ -13,6 +13,7 @@ import (
 	"github.com/UpCloudLtd/progress/messages"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/database"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/kubernetes"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/network"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/networkpeering"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands/objectstorage"
@@ -28,6 +29,7 @@ const (
 	includeHelp = "Include resources matching the given name. If defined multiple times, resource is included if it matches any of the given names. `*` matches all resources."
 	excludeHelp = "Exclude resources matching the given name. If defined multiple times, resource is included if it matches any of the given names."
 
+	typeKubernetes     = "kubernetes-cluster"
 	typeNetwork        = "network"
 	typeNetworkPeering = "network-peering"
 	typeRouter         = "router"
@@ -144,10 +146,11 @@ func findResources[T any](exec commands.Executor, wg *sync.WaitGroup, returnChan
 
 func listResources(exec commands.Executor, include, exclude []string) ([]Resource, error) {
 	var resources []Resource
-	returnChan := make(chan findResult, 9)
+	returnChan := make(chan findResult, 10)
 
 	var wg sync.WaitGroup
 
+	findResources(exec, &wg, returnChan, &resolver.CachingKubernetes{}, include, exclude)
 	findResources(exec, &wg, returnChan, &resolver.CachingNetwork{}, include, exclude)
 	findResources(exec, &wg, returnChan, &resolver.CachingNetworkPeering{}, include, exclude)
 	findResources(exec, &wg, returnChan, &resolver.CachingRouter{Type: "normal"}, include, exclude)
@@ -181,6 +184,12 @@ func listResources(exec commands.Executor, include, exclude []string) ([]Resourc
 
 func getResource(val any) (Resource, error) {
 	switch v := val.(type) {
+	case upcloud.KubernetesCluster:
+		return Resource{
+			Name: v.Name,
+			Type: typeKubernetes,
+			UUID: v.UUID,
+		}, nil
 	case upcloud.Network:
 		return Resource{
 			Name: v.Name,
@@ -241,6 +250,8 @@ func getResource(val any) (Resource, error) {
 
 func deleteResource(exec commands.Executor, resource Resource) (err error) {
 	switch resource.Type {
+	case typeKubernetes:
+		_, err = kubernetes.Delete(exec, resource.UUID, true)
 	case typeNetwork:
 		_, err = network.Delete(exec, resource.UUID)
 	case typeNetworkPeering: